Create update-user form using material UI dialog Relates #133
